--- src/components/Banner/Banner.module.scss
@@ -1,36 +1,0 @@
-@import '../../App.scss';
-
-.container {
-  position: relative;
-  overflow: hidden;
-}
-
-.slide {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  gap: 25px;
-}
-
-.image {
-  height: 320px;
-  width: 100%;
-  object-fit: cover;
-
-  @include respond-to(tablet) {
-    height: 189px;
-  }
-
-  @include respond-to(desktop) {
-    height: 400px;
-  }
-}
-
-
-
-
-
-
-
-
