+++ src/types/Banner.ts
@@ -1,0 +1,7 @@
+export interface Banner  {
+  imageUrl: string,
+  altText: string,
+  model: string,
+  title: string,
+  way: string,
+}
